requirements.txt
numpy==1.24.2

PyMySQL

python-dotenv

jupyterlab

matplotlib

pandas

Flask

connexion[swagger-ui]==2.14.1




